Extract datastore module to share DataStore<Preferences>
